Add effect size specification
Specify distance between the two classes using the mahalanobis distance. This is used to calculate a factor to multiply the betas by, essentially scaling up or down the prominence of the injected pattern. This is nicely validated in simulation, where we can see a correspondance between the effect size and the decoding accuracy