BF: divided effect size
The centroids don't sit on each ends of the random vector. Rather, the vector gets multiplied by 1 and -1, so the centroids of each class sit on two ends of a vector that's twice the length of the random vector. So we need to divide the effect size by two to account for it